# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Azure STIG Dashboard API
  description: |
    RESTful API for the Azure STIG Dashboard.
    Provides machine inventory, compliance findings, scan management, and STIG checklist export.
  version: 1.0.0
  contact:
    name: Azure STIG Dashboard
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
paths:
...
  /stigs/release-candidates/{candidateId}/apply:
    post:
      summary: Approve and activate the exact hashed package for a staged release
      tags: [STIGs]
      parameters:
        - { name: candidateId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '201': { description: Release activated }
        '400': { description: Invalid candidate ID }
        '403': { description: STIG import permission required }
...
  /stigs/import:
    post:
      summary: Trigger STIG content import from DISA
      tags: [STIGs]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                benchmarkTitles:
                  type: array
                  items: { type: string }
                  description: Specific benchmark titles to import (omit for all defaults)
                force:
                  type: boolean
                  default: false
                  description: Re-import even if version is already active
                dryRun:
                  type: boolean
                  default: false
      responses:
        '202': { description: Import started }
        '401': { description: Unauthorized }
        '403': { description: Forbidden }
  /stigs/update-check:
    post:
      summary: Check DISA catalog for newer STIG versions
      tags: [STIGs]
      security:
        - bearerAuth: []
      responses:
        '202': { description: Check started }
        '409': { description: Check already in progress }
...
  /stigs/{benchmarkId}:
    get:
      summary: Get benchmark details and version history
      tags: [STIGs]
      parameters:
        - { name: benchmarkId, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
  /stigs/{benchmarkId}/controls:
    get:
      summary: List controls for a STIG benchmark
      tags: [STIGs]
      parameters:
        - { name: benchmarkId, in: path, required: true, schema: { type: string } }
        - { name: version, in: query, schema: { type: string }, description: 'e.g. V2R8 (defaults to active)' }
        - { name: severity, in: query, schema: { type: string, enum: [high, medium, low] } }
        - { name: checkType, in: query, schema: { type: string } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 100 } }
      responses:
        '200': { description: OK }
        '404': { description: Benchmark or version not found }
  /stigs/{benchmarkId}/scan:
    post:
      summary: Assess an operating-system STIG on applicable active machines
      tags: [STIGs]
      security:
        - bearerAuth: []
      parameters:
        - { name: benchmarkId, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                machineIds:
                  type: array
                  items: { type: string }
                  description: Specific active machine IDs to assess (omit for all applicable machines)
                version:
                  type: string
                  description: 'STIG version to use e.g. V2R8 (defaults to active)'
      responses:
        '202': { description: Scan triggered }
        '404': { description: Benchmark not found }
        '422': { description: No active machines are applicable to this operating-system STIG }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Azure AD JWT token obtained via MSAL
...